# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CheeseTypeAPI",
    "version": "v1"
  },
  "paths": {
    "/api/CheeseType": {
      "get": {
        "tags": [
          "CheeseType"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CheeseType"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CheeseType"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CheeseType"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CheeseType"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheeseType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CheeseType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CheeseType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CheeseType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheeseType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheeseType"
                }
              }
            }
          }
        }
      }
    },
    "/api/CheeseType/{id}": {
      "get": {
        "tags": [
          "CheeseType"
        ],
        "operationId": "GetCheeseType",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CheeseType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheeseType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheeseType"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CheeseType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheeseType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CheeseType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CheeseType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "CheeseType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CheeseType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      }
    }
  }
}